move message handling and image processing logic out of tools to askbot
