+++ tools.py
@@ -1,16 +1,13 @@
 import asyncio
-import base64
 import inspect
 import logging
 import os
-from collections import OrderedDict
 from datetime import datetime
 from zoneinfo import ZoneInfo
 
 import aiohttp
 import discord
 import regex
-from bs4 import BeautifulSoup
 from ollama import Client
 
 import utils
@@ -78,9 +75,6 @@ class AITools:
 
         self.model = "artifish/llama3.2-uncensored"
         self.vision_model = "llava"
-
-        self.message_cache = OrderedDict()
-        self.cache_size = 1000
 
         self.ollama_endpoint = os.getenv("OLLAMA_ENDPOINT")
         self.searxng_endpoint = os.getenv("SEARXNG_ENDPOINT")
@@ -162,7 +156,7 @@ class AITools:
                 response = await asyncio.to_thread(
                     self.client.chat,
                     model=model,
-                    messages=messages,
+                    messages=[system_instructions] + list(messages),
                     tools=self.tool_definitions,
                 )
             except Exception:
@@ -260,181 +254,6 @@ class AITools:
                 }
 
         return await asyncio.gather(*(_run_tool(call) for call in tool_calls))
-
-    async def _populate_messages(self, payload):
-        """
-        Creates a message history from a single message,
-        looking through all of it's replies
-        :param payload: Message to start search from
-        :return: Message history
-        """
-        messages = []
-        current = await payload.channel.fetch_message(payload.reference.message_id)
-
-        while current:
-            image_urls = await self._extract_image_urls(current)
-            images_b64 = set()
-            if image_urls:
-                for url in image_urls:
-                    images_b64.add(await self._url_to_base64(url))
-
-            messages.append(
-                {
-                    "role": "assistant" if current.author.bot else "user",
-                    "content": regex.sub(
-                        r"<@!?(\d+)>",
-                        lambda m: (
-                            (current.guild.get_member(int(m.author.id))).name
-                            if payload.guild.get_member(int(m.author.id))
-                            else m.group(0)
-                        ),
-                        current.content,
-                    ),
-                    "images": images_b64 if images_b64 else "",
-                }
-            )
-
-            if current.reference:
-                current = await self._get_message(
-                    current.channel, current.reference.message_id
-                )
-
-            else:
-                break
-
-        messages.reverse()
-
-        image_urls = await self._extract_image_urls(payload)
-        images_b64 = set()
-        if image_urls:
-            for url in image_urls:
-                images_b64.add(await self._url_to_base64(url))
-        messages.append(
-            {
-                "role": "user",
-                "content": payload.content,
-                "images": images_b64 if images_b64 else "",
-            }
-        )
-
-        return list(messages)
-
-    async def _get_message(self, channel: discord.TextChannel, message_id: int):
-        """
-        Helper function to find a message object, looks for a cached message
-        first, falling back to the discord API if not found.
-        :param channel: Channel to look in
-        :param message_id: Message ID to look for
-        :return: Message object
-        """
-        if message_id in self.message_cache:
-            return self.message_cache[message_id]
-
-        try:
-            message = await channel.fetch_message(message_id)
-            await self._cache_message(message_id, message)
-            return message
-
-        except (discord.NotFound, discord.Forbidden) as e:
-            self.logger.error(f"FAILED TO GET MESSAGE: {e}")
-
-    async def _cache_message(self, message_id, message):
-        """
-        Cache a message object
-        :param message_id: message ID to cache
-        :param message: message object to cache
-        :return:
-        """
-        self.message_cache[message_id] = message
-        if len(self.message_cache) > self.cache_size:
-            self.message_cache.popitem(last=False)
-
-    async def _url_to_base64(self, url: str) -> str:
-        """
-        Converts an image URL to a base64 encoded string.
-        :param url: Image URL
-        :return: Base64 encoded string
-        """
-        try:
-            async with aiohttp.ClientSession() as session:
-                tries = 0
-
-                while url and tries < 5:
-                    async with session.get(url, timeout=10) as response:
-                        content_type = response.headers.get("Content-Type", "")
-                        if "image" in content_type:
-                            data = await response.read()
-                            return base64.b64encode(data).decode("utf-8")
-
-                        elif "text/html" in content_type:
-                            self.logger.debug(f"HTML found: {url}")
-                            html = await response.content.read()
-                            soup = BeautifulSoup(html, "html.parser")
-
-                            og_image = soup.find("meta", property="og:image")
-                            if og_image and og_image.get("content"):
-                                url = og_image["content"]
-                                tries += 1
-                                continue
-
-                        else:
-                            self.logger.debug(f"NO IMAGE IN URL: {url}")
-                            return None
-
-        except aiohttp.ClientError as e:
-            self.logger.error(f"FAILED TO FETCH URL {url}: {e}")
-            return None
-
-    async def _extract_image_urls(self, message: discord.Message):
-        """
-        Extracts image URLs from a message and stores them in a set.
-        :param message: Message object to extract image URLs from
-        :return: Image URLs
-        """
-        image_urls = set()
-
-        for attachment in message.attachments:
-            if attachment.content_type == "image":
-                image_urls.add(attachment.url)
-
-        for embed in message.embeds:
-            if embed.thumbnail.url:
-                image_urls.add(embed.thumbnail.url)
-            if embed.image.url:
-                image_urls.add(embed.image.url)
-
-        urls = regex.findall(
-            r"https?://[^\s]+", message.content, flags=regex.IGNORECASE
-        )
-        for url in urls:
-            if url.lower().endswith((".jpg", ".jpeg", ".png", ".gif")):
-                image_urls.add(url)
-
-            else:
-                try:
-                    async with aiohttp.ClientSession() as session:
-                        async with session.get(url, timeout=10) as response:
-                            content_type = response.headers.get("Content-Type", "")
-                            if "text/html" in content_type:
-                                html = await response.text()
-                                soup = BeautifulSoup(html, "html.parser")
-
-                                # og:image
-                                og_image = soup.find("meta", property="og:image")
-                                if og_image and og_image.get("content"):
-                                    image_urls.add(og_image["content"])
-
-                                elif (img := soup.find("img")) and img.get("src"):
-                                    image_urls.add(img["src"])
-
-                except aiohttp.ClientError:
-                    self.logger.exception(f"Failed to fetch HTML page {url}")
-
-        if not image_urls:
-            self.logger.debug("NO IMAGE URLS FOUND")
-            return None
-
-        return image_urls
 
     @tool
     def _respond_to_user(self, response=None) -> str:
